chore(benchmark): modularized high-level functions
